rename: `table.ts` -> `ui.ts`로 변경 (#6)
