+++ components/common/TableCell/index.tsx
@@ -1,5 +1,5 @@
 import * as styles from '@/styles/table.css'
-import { TableCellProps } from '@/types/table'
+import { TableCellProps } from '@/types/ui'
 
 const TableCell = ({ children, colSpan, rowSpan, className }: TableCellProps) => (
     <td colSpan={colSpan} rowSpan={rowSpan} className={`${styles.tableCell} ${className || ''}`}>
