+++ components/common/TableHeader/index.tsx
@@ -1,5 +1,5 @@
 import * as styles from '@/styles/table.css'
-import { TableHeaderProps } from '@/types/table'
+import { TableHeaderProps } from '@/types/ui'
 
 const TableHeader = ({ children, scope = 'col', rowSpan, colSpan, className }: TableHeaderProps) => {
     return (
